working on Dockerfile, need to reorganize file structure for exstension conversion
